Output page skeleton
